CIRCSTORE-407 Merge CirculationStorageKafkaTopic and InventoryKafkaTopic
