+++ src/main/java/org/folio/service/event/InventoryEventType.java
@@ -1,6 +1,6 @@
 package org.folio.service.event;
 
-import static org.folio.support.kafka.topic.InventoryKafkaTopic.ITEM;
+import static org.folio.FolioKafkaTopic.ITEM;
 import static org.folio.service.event.InventoryEventType.PayloadType.UPDATE;
 
 import org.folio.kafka.services.KafkaTopic;
